[NPU] fixed unit-tests compilation under sanitizer  (#31795)
### Details:
- limit npu-unit-test source code to only required .cpp files, to fix
issue with linkage under openvino -DENABLE_SANITIZER=ON
### Tickets:
 - *E-180030*
